server/requirements.txt
flask
flask-cors
ultralytics
torch
numpy
pillow